... ah thank you (mark wild) so much...
ok i am rotating on the y first and that is a offset left or right exactly on the y system axis which is not at all what i should be doing.
I should be rotating the normal to form a offset rotating about the X axis to offset the vector up or down.
Thats step one and i might be able to form a proper oriention as well to avoid gimble singularitys like that.